requirements.txt
sentence-transformers==5.1.1
numpy==2.3.3
nltk==3.9.1
faiss-cpu==1.12.0
